requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=2.1.0=pypi_0
aiohappyeyeballs=2.4.0=pypi_0
aiohttp=3.10.5=pypi_0
aiosignal=1.3.1=pypi_0
async-timeout=4.0.3=pypi_0
attrs=24.2.0=pypi_0
blas=1.0=mkl
brotli-python=1.1.0=py310h9e98ed7_2
bzip2=1.0.8=h2466b09_7
ca-certificates=2024.8.30=h56e8100_0
cachetools=5.5.0=pypi_0
certifi=2024.8.30=pyhd8ed1ab_0
cffi=1.17.0=py310ha8f682b_1
charset-normalizer=3.3.2=pyhd8ed1ab_0
cloudpickle=3.0.0=pypi_0
colorama=0.4.6=pypi_0
cuda-cccl=12.6.37=0
cuda-cccl_win-64=12.6.37=0
cuda-cudart=11.8.89=0
cuda-cudart-dev=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-libraries-dev=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvrtc-dev=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-profiler-api=12.6.68=0
cuda-runtime=11.8.0=0
cuda-version=12.6=3
datasets=2.21.0=pypi_0
dill=0.3.8=pypi_0
farama-notifications=0.0.4=pypi_0
filelock=3.15.4=pyhd8ed1ab_0
freetype=2.12.1=hdaf720e_2
frozenlist=1.4.1=pypi_0
fsspec=2024.6.1=pypi_0
geoparse=2.0.4=pypi_0
google-auth=2.34.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
grpcio=1.66.1=pypi_0
gymnasium=0.28.1=pypi_0
h2=4.1.0=pyhd8ed1ab_0
hpack=4.0.0=pyh9f0ad1d_0
huggingface-hub=0.24.6=pypi_0
hyperframe=6.0.1=pyhd8ed1ab_0
idna=3.8=pyhd8ed1ab_0
imbalanced-learn=0.12.3=pypi_0
intel-openmp=2024.2.1=h57928b3_1083
jax-jumpy=1.0.0=pypi_0
jinja2=3.1.4=pyhd8ed1ab_0
joblib=1.4.2=pypi_0
lcms2=2.16=h67d730c_0
lerc=4.0.0=h63175ca_0
libblas=3.9.0=1_h8933c1f_netlib
libcblas=3.9.0=5_hd5c7e75_netlib
libcublas=11.11.3.6=0
libcublas-dev=11.11.3.6=0
libcufft=10.9.0.58=0
libcufft-dev=10.9.0.58=0
libcurand=10.3.7.68=0
libcurand-dev=10.3.7.68=0
libcusolver=11.4.1.48=0
libcusolver-dev=11.4.1.48=0
libcusparse=11.7.5.86=0
libcusparse-dev=11.7.5.86=0
libdeflate=1.21=h2466b09_0
libffi=3.4.2=h8ffe710_5
libhwloc=2.11.1=default_h8125262_1000
libiconv=1.17=hcfcfb64_2
libjpeg-turbo=3.0.0=hcfcfb64_1
liblapack=3.9.0=5_hd5c7e75_netlib
libnpp=11.8.0.86=0
libnpp-dev=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libnvjpeg-dev=11.9.0.86=0
libpng=1.6.43=h19919ed_0
libsqlite=3.46.1=h2466b09_0
libtiff=4.6.0=hb151862_4
libuv=1.48.0=hcfcfb64_0
libwebp-base=1.4.0=hcfcfb64_0
libxcb=1.16=h013a479_1
libxml2=2.12.7=h0f24e4e_4
libzlib=1.3.1=h2466b09_1
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown=3.7=pypi_0
markupsafe=2.1.5=py310ha8f682b_1
mkl=2023.1.0=h6a75c08_48682
mpmath=1.3.0=pyhd8ed1ab_0
msys2-conda-epoch=20160418=1
multidict=6.1.0=pypi_0
multiprocess=0.70.16=pypi_0
networkx=3.3=pyhd8ed1ab_1
numpy=1.24.3=pypi_0
oauthlib=3.2.2=pypi_0
openjpeg=2.5.2=h3d672ee_0
openssl=3.3.2=h2466b09_0
packaging=24.1=pypi_0
pandas=2.0.3=pypi_0
pillow=10.4.0=py310h3e38d90_0
pip=24.2=pyh8b19718_1
pmlb=1.0.1.post3=pypi_0
protobuf=5.28.0=pypi_0
pthread-stubs=0.4=hcd874cb_1001
pthreads-win32=2.9.1=hfa6e2cd_3
pyarrow=17.0.0=pypi_0
pyasn1=0.6.0=pypi_0
pyasn1-modules=0.4.0=pypi_0
pycparser=2.22=pyhd8ed1ab_0
pysocks=1.7.1=pyh0701188_6
python=3.10.14=h4de0772_0_cpython
python_abi=3.10=5_cp310
pytorch=2.4.1=py3.10_cuda11.8_cudnn9_0
pytorch-cuda=11.8=h24eeafa_5
pytorch-mutex=1.0=cuda
pytz=2024.1=pypi_0
pyyaml=6.0.2=py310ha8f682b_1
requests=2.32.3=pyhd8ed1ab_0
requests-oauthlib=2.0.0=pypi_0
rsa=4.9=pypi_0
scikit-learn=1.3.0=pypi_0
scipy=1.14.1=pypi_0
setuptools=73.0.1=pyhd8ed1ab_0
sympy=1.13.2=pyh04b8f61_3
tbb=2021.12.0=hc790b64_4
tensorboard=2.12.3=pypi_0
tensorboard-data-server=0.7.2=pypi_0
threadpoolctl=3.5.0=pypi_0
tk=8.6.13=h5226925_1
torchaudio=2.4.1=pypi_0
torchvision=0.19.1=pypi_0
tqdm=4.66.5=pypi_0
typing_extensions=4.12.2=pyha770c72_0
tzdata=2024.1=pypi_0
ucrt=10.0.22621.0=h57928b3_0
urllib3=2.2.2=pyhd8ed1ab_1
vc=14.3=h8a93ad2_20
vc14_runtime=14.40.33810=hcc2c482_20
vs2015_runtime=14.40.33810=h3bf8584_20
werkzeug=3.0.4=pypi_0
wheel=0.44.0=pyhd8ed1ab_0
win_inet_pton=1.1.0=pyhd8ed1ab_6
xorg-libxau=1.0.11=hcd874cb_0
xorg-libxdmcp=1.1.3=hcd874cb_0
xxhash=3.5.0=pypi_0
xz=5.2.6=h8d14728_0
yaml=0.2.5=h8ffe710_2
yarl=1.11.1=pypi_0
zstandard=0.23.0=py310he5e10e1_1
zstd=1.5.6=h0ea2cb4_0
